Test tweaks
* use the fixtures helper in more places
* replace inline fixtures() calls with a hoisted const
* convert assert.ok() to assert.equal
* split tests
