+++ package.json
@@ -97,6 +97,14 @@
       "unicorn/prefer-node-protocol": "off",
       "unicorn/prefer-top-level-await": "off",
       "unicorn/prevent-abbreviations": "off"
-    }
+    },
+    "overrides": [
+      {
+        "files": "test/**/*.js",
+        "envs": [
+          "mocha"
+        ]
+      }
+    ]
   }
 }
